hide blueprint lock icons on child courses
A front-end performance hack on courses with large numbers of modules.
If this feature flag is on we hide the lock icons on child courses and
instead append a lock emoji to the title when there are restrictions.
This is a performance hack that should be sparingly used.

flag=modules_page_hide_blueprint_lock_icon_for_children

test plan:
  - turn on the feature flag
  - Create a blueprint parent course with an assignment that has some
    restrictions
  - Create a child course from the blueprint
  - Go to the modules page in the child course and verify that the lock
    icon is not added via javascript (on the right side of the page) but
    is an emoji added from the ERB near the title

Change-Id: I968d92396d0b099d121f71360a1ea97c3dd4c9cf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/356281
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>